requirements.txt
asyncpg==0.30.0
fastapi==0.116.1
httpx==0.28.1
passlib==1.7.4
pydantic==2.11.9
pydantic_settings==2.10.1
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
python_jose==3.5.0
slowapi==0.1.9
supabase==2.18.1
uvicorn==0.23.2
google-generativeai==0.3.2